Fix implicit conversion of returned value in AssetLibrary::numEffects()

+++ framework/include/minko/file/AssetLibrary.hpp
@@ -106,7 +106,7 @@ namespace minko
             uint
 			numEffects()
 			{
-				return _effects.size();
+				return static_cast<uint>(_effects.size());
 			}
 
 			inline
